protocol: fixed missing field of CraftRecipeAuto
