--- src/pocketmine/network/mcpe/protocol/types/inventory/stackrequest/CraftRecipeStackRequestActionTrait.php
@@ -1,47 +1,0 @@
-<?php
-
-/*
- *
- *  ____            _        _   __  __ _                  __  __ ____
- * |  _ \ ___   ___| | _____| |_|  \/  (_)_ __   ___      |  \/  |  _ \
- * | |_) / _ \ / __| |/ / _ \ __| |\/| | | '_ \ / _ \_____| |\/| | |_) |
- * |  __/ (_) | (__|   <  __/ |_| |  | | | | | |  __/_____| |  | |  __/
- * |_|   \___/ \___|_|\_\___|\__|_|  |_|_|_| |_|\___|     |_|  |_|_|
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Lesser General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * @author PocketMine Team
- * @link http://www.pocketmine.net/
- *
- *
-*/
-
-declare(strict_types=1);
-
-namespace pocketmine\network\mcpe\protocol\types\inventory\stackrequest;
-
-use pocketmine\network\mcpe\NetworkBinaryStream;
-
-trait CraftRecipeStackRequestActionTrait{
-
-	/** @var int */
-	private $recipeId;
-
-	final public function __construct(int $recipeId){
-		$this->recipeId = $recipeId;
-	}
-
-	public function getRecipeId() : int{ return $this->recipeId; }
-
-	public static function read(NetworkBinaryStream $in) : self{
-		$recipeId = $in->readGenericTypeNetworkId();
-		return new self($recipeId);
-	}
-
-	public function write(NetworkBinaryStream $out) : void{
-		$out->writeGenericTypeNetworkId($this->recipeId);
-	}
-}
